FOLDER REORGANIZATION
Did some more folder renaming, refactoring, and organization to make the code cleaner in the file viewer--if this really does that. The idea is that the scenes themselves are
model, whereas the guis are the view parts of it. Then there's a scene controller that manages them.
