+++ src/view/guiparts/buttons/GameButtons.java
@@ -1,0 +1,4 @@
+package view.guiparts.buttons;
+
+public enum GameButtons {
+}
