+++ src/view/guis/MainMenuGUI.java
@@ -1,8 +1,7 @@
-package view.scenes;
+package view.guis;
 
 import controller.MainMenuController;
-import view.parts.MainMenuButtonPanel;
-import view.parts.TextLog;
+import view.guiparts.MainMenuButtonPanel;
 
 import javax.swing.*;
 
@@ -14,8 +13,7 @@ public class MainMenuGUI extends GUI {
     public MainMenuGUI(MainMenuController mmc){
         super(new JLabel("RogueCrawler Timescapes"),new MainMenuButtonPanel(mmc));
 
-        // Load in an image -- make something simple!
-
+        // TODO: Load in an image in the center so buttons go to the bottom
 
 
     }
