--- src/view/parts/buttons/GameButtons.java
@@ -1,4 +1,0 @@
-package view.parts.buttons;
-
-public enum GameButtons {
-}
